requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
attrs==17.4.0
autobahn==18.4.1
Automat==0.6.0
bcrypt==3.1.4
cffi==1.9.1
channels==1.1.8
constantly==15.1.0
cryptography==2.1.4
daphne==1.4.2
Django==1.11.6
djangorestframework==3.6.4
djangorestframework-yaml==1.0.3
enum34==1.1.6
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.18
msgpack-python==0.5.6
olefile==0.44
paramiko==2.4.0
Pillow==5.0.0
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
python-magic==0.4.3
pytz==2018.3
PyYAML==3.12
redis==2.10.6
six==1.11.0
Twisted==17.9.0
txaio==2.10.0
zope.interface==4.4.3
ansible==2.4.2.0
flower==0.9.2
django-celery==3.2.2
redis==2.10.6
django-jet==1.0.7
xlrd==1.1.0
xlwt==1.3.0
chardet==3.0.4
feedparser==5.2.1
